Fix OTP authentication class exhausting the stream
Apparently, `rest_framework.request.clone_request()` uses the same
socket stream for the cloned request as the original.  This means that
if any stacked authenticator class does anything that reads from the
socket, it's taken out from under the original request.

Switch to an approach that doesn't use `clone_request()`.